- Scripting system undergone some additions and changes - System-Global UI added (now referred to as Global mostly) to handle overlay effects and later on error-dialogs, etc.

+++ src/Script/ScriptManager.h
@@ -30,7 +30,7 @@ public:
 		- Alternatively a boolean could be set which toggles this behaviour, but plan ahead and make sure that whatever you do doesn't leak memory!
 	*/
 	void PlayScript(Script * script);
-	void Process(float timeInSeconds);
+	void Process(long long timeInSeconds);
 	List<Script*> GetActiveEvents(){ return activeEvents; };
 private:
 	List<Script*> activeEvents;
